\name{GOSPEL-package}
\alias{GOSPEL-package}
\alias{GOSPEL}
\docType{package}
\title{
\packageTitle{GOSPEL}
}
\description{
\packageDescription{GOSPEL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GOSPEL}
\packageIndices{GOSPEL}
We propose a sparse learning algorithm for network graph data, called Graph-Oriented SParcE Learning (GOSPEL), to find a subset of the topological information matrices of the predictor variables (networks) related to the response variable (network). More specifically, we propose to use particular forms of diffusion kernel-based centered kernel alignment as a measure of statistical correlation between graph Laplacian matrices, and solve the optimization problem with a novel graph-guided generalized fused lasso.
}
\author{
\packageAuthor{GOSPEL}

Maintainer: \packageMaintainer{GOSPEL}
}
\references{
Hideko Kawakubo, Yusuke Matsui, Itaru Kushima, Norio Ozaki, and Teppei Shimamura, A Network of Networks Approach for Modeling Interconnected Brain Tissue-Specific Networks, submitted.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
