\name{GOSPEL-package}
\alias{GOSPEL-package}
\alias{GOSPEL}
\docType{package}
\title{
\packageTitle{GOSPEL}
}
\description{
\packageDescription{GOSPEL}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{GOSPEL}
\packageIndices{GOSPEL}
}
\author{
\packageAuthor{GOSPEL}
Maintainer: \packageMaintainer{GOSPEL}
}
\references{
A Network of Networks Approach for Modeling Interconnected Brain Tissue-Specific Network
}
